requirements.txt
requests~=2.31.0

typing_extensions~=4.12.2
PyGithub~=2.3.0